Feat: 이미지와 비디오 업로드 기능 코드 통합
